Support for Extra Double Quotes in GraphViz files
When we convert ndfa or dfa to GraphViz generally extra double quotes are observed and therefore cannot be run on GraphViz. This is a simple method to scan and remove all these extra quotes and the dot file so generated can be run using gvedit. Hope this helps many others who would have suffered from the same problem :)